requirements.txt
openai>=1.109.1,<3.0.0

langchain==1.0.5
langchain-core>=1.2.6
langchain-openai>=1.1.0
langgraph==1.0.3
langsmith==0.3.45

python-dotenv>=1.0.0